Introduced `PLL_Widgets_Trait`
